+++ NEWS.md
@@ -3,6 +3,7 @@
 * Add option to use GET in sendQuery internal function
 * getTestStatus checks for statusCode <= 200L now
 * Filter argument added to getHistory
+* Add getPageSpeedData function, tests and docs
 
 # 0.0.3
 * Add getTesters function, tests and docs